open_RGB/requirements.txt
openrgb-python>=0.2.16
numpy>=1.21.0
websockets>=11.0.0
python-dotenv>=1.0.0
